+++ api/v1beta1/dataplane_types.go
@@ -34,7 +34,7 @@ func init() {
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:shortName=kodp,categories=kong;all
-// +kubebuilder:validation:XValidation:message="Extension not allowed for DataPlane",rule="has(self.spec.extensions) ? self.spec.extensions.all(e, e.group == 'gateway-operator.konghq.com' && e.kind == 'DataPlaneKonnectExtension') : true"
+// +kubebuilder:validation:XValidation:message="Extension not allowed for DataPlane",rule="has(self.spec.extensions) ? self.spec.extensions.all(e, e.group == 'gateway-operator.konghq.com' && e.kind == 'KonnectExtension') : true"
 // +kubebuilder:printcolumn:name="Ready",description="The Resource is ready",type=string,JSONPath=`.status.conditions[?(@.type=='Ready')].status`
 
 // DataPlane is the Schema for the dataplanes API
@@ -78,7 +78,7 @@ type DataPlaneOptions struct {
 
 	// Extensions provide additional or replacement features for the DataPlane
 	// resources to influence or enhance functionality.
-	// NOTE: since we have one extension only (DataPlaneKonnectExtension), we limit the amount of extensions to 1.
+	// NOTE: since we have one extension only (KonnectExtension), we limit the amount of extensions to 1.
 	//
 	// +optional
 	// +kubebuilder:validation:MinItems=0