+++ controller/dataplane/dataplanekonnectextension_rbac.go
@@ -1,9 +1,0 @@
-package dataplane
-
-// -----------------------------------------------------------------------------
-// DataPlaneKonnectExtensionReconciler - RBAC
-// -----------------------------------------------------------------------------
-
-// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=dataplanes,verbs=get;list;watch
-// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=dataplanekonnectextensions,verbs=get;list;watch
-// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=dataplanekonnectextensions/finalizers,verbs=update;patch