+++ controller/dataplane/konnectextension_controller_rbac.go
@@ -1,0 +1,9 @@
+package dataplane
+
+// -----------------------------------------------------------------------------
+// KonnectExtensionReconciler - RBAC
+// -----------------------------------------------------------------------------
+
+// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=dataplanes,verbs=get;list;watch
+// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=konnectextensions,verbs=get;list;watch
+// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=konnectextensions/finalizers,verbs=update;patch